+++ packages/axios/src/index.ts
@@ -1,11 +1,8 @@
 import { computed, Ref, ref } from "@vue/composition-api";
 import axios, { AxiosRequestConfig, AxiosResponse, AxiosInstance, CancelTokenSource } from "axios";
-import { usePromise, PromiseResultFactory } from "@vue-composable/core";
+import { usePromise, PromiseResultFactory, isString, isBoolean, isObject } from "@vue-composable/core";
 
-/* istanbul ignore next  */
-const _axios = axios || (globalThis && (globalThis as any).axios);
-
-interface AxiosReturn<TData> extends PromiseResultFactory<Promise<AxiosResponse<TData>>, [AxiosRequestConfig]> {
+interface AxiosReturn<TData> extends PromiseResultFactory<Promise<AxiosResponse<TData>>, [AxiosRequestConfig | string]> {
   readonly client: Ref<Readonly<AxiosInstance>>;
   readonly data: Ref<TData | null>;
   readonly status: Ref<number | null>;
@@ -17,12 +14,6 @@ interface AxiosReturn<TData> extends PromiseResultFactory<Promise<AxiosResponse<
   // readonly 
 }
 
-// TODO replace by shared project
-const isObject = (d: any): d is Object => typeof d === 'object';
-const isString = (d: any): d is String => typeof d === 'string';
-const isBoolean = (d: any): d is Boolean => typeof d === 'boolean';
-
-
 export function useAxios<TData = any>(throwException?: boolean): AxiosReturn<TData>;
 export function useAxios<TData = any>(url: string, config?: AxiosRequestConfig, throwException?: boolean): AxiosReturn<TData>;
 export function useAxios<TData = any>(url: string, throwException?: boolean): AxiosReturn<TData>;
@@ -30,13 +21,13 @@ export function useAxios<TData = any>(config?: AxiosRequestConfig, throwExceptio
 export function useAxios<TData = any>(configUrlThrowException?: AxiosRequestConfig | string | boolean, configThrowException?: AxiosRequestConfig | boolean, throwException = false): AxiosReturn<TData> {
   /* istanbul ignore next  */
   __DEV__ &&
-    !_axios &&
+    !axios &&
     console.warn(`[axios] not installed, please install it`);
 
   const config = !isString(configUrlThrowException) && !isBoolean(configUrlThrowException) ? configUrlThrowException : isObject(configThrowException) ? configThrowException as AxiosRequestConfig : undefined;
   throwException = isBoolean(configUrlThrowException) ? configUrlThrowException : isBoolean(configThrowException) ? configThrowException : throwException;
 
-  const axiosClient = _axios.create(config);
+  const axiosClient = axios.create(config);
   const client = computed(() => axiosClient);
   const isCancelled = ref(false);
   const cancelledMessage = ref<string>(null);
@@ -54,14 +45,16 @@ export function useAxios<TData = any>(configUrlThrowException?: AxiosRequestConf
     }
   }
 
-  const use = usePromise(async (request: AxiosRequestConfig) => {
-    cancelToken = _axios.CancelToken.source()
+  const use = usePromise(async (request: AxiosRequestConfig | string) => {
+    cancelToken = axios.CancelToken.source()
     isCancelled.value = false;
     cancelledMessage.value = null;
 
+    const opts = isString(request) ? { url: request } : request;
+
     return axiosClient.request<any, AxiosResponse<TData>>({
       cancelToken: cancelToken.token,
-      ...request
+      ...opts
     });
   }, throwException);