+++ packages/web/__tests__/storage/storage.spec.ts
@@ -1,0 +1,9 @@
+import { useStorage } from '../../src';
+
+describe('storage', () => {
+  it('should get localStorage if supported', () => {
+    const { storage } = useStorage('');
+
+    expect(storage).not.toBeNull();
+  })
+})