+++ packages/web/src/web/index.ts
@@ -2,4 +2,7 @@ export * from "./fetch";
 export * from "./webSocket";
 export * from "./intersectionObserver";
 export * from "./networkInformation";
+export * from "./online";
+export * from "./pageVisibility";
+export * from "./language";
 export * from "./geolocation";